feat: user feedback - Selection and different card sfx
The player can now experience feedback on selection of cards, units, tiles and the like. A unique sound effect is also played for each card based on the name of said card. Defaults to basic slash if it can't find a name that matches. 

Added a getter function to card to facilitate the logic.
Added clean up