feat: implement correct player designation and ability to draw
The scene is passed a boolean on whether the client is initially a sender of receiver from the interface class. This designates whether they are 'O' or 'X'. Make use of STL via std::all_of to discover a draw in which all tiles have been filled but no win condition has been met yet.